Have TWO-WAY- and ECHO-MISC forward more carefully and identically.
1) Forward :CHARPOS or :LINE-LENGTH only to the output
component. (Both two-way and echo streams have forwarded these to the
input component.)

2) Forward :CLEAR-INPUT only to input streams. (This is not a change
for two-way streams. Echo streams used to forward to the output
component.)

3) Forward :CLEAR-OUTPUT, :FINISH-OUTPUT, :FORCE-OUTPUT only to the
output component. (This is not a change for two-way streams. Echo
streams used to forward to the output component.)

4) Make the two functions as textually-identical as seemed possible &
prudent.

(1) through (3) make straightforwardly-implemented, unidirectional
Gray Streams as well as bidirectional streams of any class more
suitable for use as components of two-way and echo streams.