Add N-CIN slot to ANSI-STREAM, methods for all constructed streams.
The slot is to hold a NIL or a function. When it's NIL,
ANSI-STREAM-READ-LINE and ANSI-STREAM-READ-SEQUENCE fall through to a
READ-CHAR loop. Otherwise it's a function that implements both
READ-LINE and READ-SEQUENCE into a string.

In order to simplify READ-LINE implementations, the return convention
is a union of Gray Streams' STREAM-READ-LINE and STREAM-READ-SEQUENCE.
ANSI-STREAM-READ-LINE, ANSI-STREAM-READ-LINE-FROM-FRC-BUFFER, and
a test are adjusted accordingly.

The net effect is that there are now 3 ways READ-LINE or
READ-SEQUENCE/string can go on an ANSI-STREAM: (1) copy from the input
buffer if there is one, (2) call the N-CIN method if there is one, (3)
otherwise fall through a READ-CHAR loop.

All standard composite streams get methods that forward to input
components. String input streams do READ-LINE as SUBSEQ, READ-SEQUENCE
into a string as REPLACE.

This commit also causes ANSI-STREAM-READ-SEQUENCE to try to take a
faster path for all strings, not just simple ones, which might be
considered an orthogonal concern.

Potential future work:

- bivalent and/or bidirectional FD-STREAMs;
- decide whether READ-LINE calls should produce strings with
  convenient or predicatble element types,
- sb-simple-streams: READ-SEQUENCE on composite of a
  file-simple-stream doesn't work right; its PARTIAL-FILL extension
  doesn't get forwarded to/through composites.